Add files via upload
In this project I have implemented a module for simulating (MLQ) CPU Scheduling.

To implement this I have used Visual studio as the tool and python as the programming language. I have used python, because of it's simplicity and reusability as well as it's very easy to create a user interface using python.